chore: make emission tests compile
